back/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clarifai==9.8.2
clarifai-grpc==9.8.6
click==8.1.7
cohere==4.27
contextlib2==21.6.0
dataclasses-json==0.6.1
dill==0.3.7
fastapi==0.103.1
fastavro==1.8.2
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.59.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.305
langsmith==0.0.41
llvmlite==0.41.0
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nlpcloud==1.1.44
numba==0.58.0
numexpr==2.8.7
numpy==1.25.2
openai==0.28.1
openai-whisper==20230918
openlm==0.0.5
packaging==23.1
protobuf==4.24.3
pycparser==2.21
pydantic==2.4.0
pydantic-settings==2.0.3
pydantic_core==2.10.0
pydub==0.25.1
Pygments==2.16.1
python-dotenv==1.0.0
python-multipart==0.0.6
python-rapidjson==1.11
PyYAML==6.0.1
redis==5.0.1
regex==2023.8.8
requests==2.31.0
rich==13.4.2
safetensors==0.3.3
schema==0.7.5
sniffio==1.3.0
soundfile==0.12.1
SQLAlchemy==2.0.21
sqlitedict==2.1.0
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.1
tqdm==4.64.1
transformers==4.33.3
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
